Move DoctrineOrmTestCase to tests directory
This class is only used in our own tests
